backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
pandas>=2.0.0
openpyxl>=3.1.0
langchain>=0.2.1
langchain-groq>=0.1.3
langchain-community>=0.2.1
langchain-huggingface>=0.0.3
faiss-cpu>=1.12.0
sentence-transformers>=2.7.0
